Other unfolding strategy
First predicate in every conjuction is unfolded (once) until every
branch either is a leaf or a call of some predicate. Then all the
subsequent predicates are "simplified": they are unfolded once and only
if the unfolding is deterministic (only one branch "survives" i.e. is
not a failing branch). Then the resulting conjunction is checked for
renaming and embedding, and is either folded to some ancestor node (if
it is a renaming), generalized (if it is an embedding) or the driving
process continues.